Fixes for printing off_t values
Seems the only portable and safe way to do this is with format
%lld and cast the argument to (long long).  There is no printf(3)
for off_t that isequivalent to %zd (for size_t or ssize_t);

Found on vm37 (OpenBSD 6.7).

+++ src/pmdumplog/pmdumplog.c
@@ -954,11 +954,11 @@ dumpTI(__pmContext *ctxp)
 	    printf("\t\tError: illegal timestamp value (%" FMT_INT64 " sec, %d nsec)\n",
 		tip->stamp.sec, tip->stamp.nsec);
 	if (meta_size != -1 && tip->off_meta > meta_size)
-	    printf("\t\tError: offset to meta file past end of file (%zd)\n",
-		meta_size);
+	    printf("\t\tError: offset to meta file past end of file (%lld)\n",
+		(long long)meta_size);
 	if (log_size != -1 && tip->off_data > log_size)
-	    printf("\t\tError: offset to log file past end of file (%zd)\n",
-		log_size);
+	    printf("\t\tError: offset to log file past end of file (%lld)\n",
+		(long long)log_size);
 	if (i > 0) {
 	    if (tip->stamp.sec < lastp->stamp.sec ||
 	        (tip->stamp.sec == lastp->stamp.sec &&
@@ -1031,7 +1031,7 @@ rawdump(FILE *f)
     int		sts;
 
     if ((old = ftell(f)) < 0) {
-	fprintf(stderr, "rawdump: Botch: ftell(%p) -> %zd (%s)\n", f, old, pmErrStr(-errno));
+	fprintf(stderr, "rawdump: Botch: ftell(%p) -> %lld (%s)\n", f, (long long)old, pmErrStr(-errno));
 	return;
     }
 
@@ -1040,7 +1040,7 @@ rawdump(FILE *f)
 
     while ((sts = fread(&len, 1, sizeof(len), f)) == sizeof(len)) {
 	len = ntohl(len);
-	printf("Dump ... record len: %d @ offset: %zd", len, (ftell(f) - sizeof(len)));
+	printf("Dump ... record len: %d @ offset: %lld", len, (long long)(ftell(f) - sizeof(len)));
 	len -= 2 * sizeof(len);
 	for (i = 0; i < len; i++) {
 	    check = fgetc(f);
@@ -1068,7 +1068,7 @@ rawdump(FILE *f)
     if (sts < 0)
 	printf("fread fails: %s\n", osstrerror());
     if (fseek(f, old, SEEK_SET) < 0)
-	fprintf(stderr, "Warning: fseek(..., %zd, ...) failed: %s\n", old, pmErrStr(-oserror()));
+	fprintf(stderr, "Warning: fseek(..., %lld, ...) failed: %s\n", (long long)old, pmErrStr(-oserror()));
 }
 
 static void
